Refactor the imagensdk
