+++ imagen_sdk/__init__.py
@@ -58,45 +58,14 @@ One-Line Usage:
     print(f"Exported files: {result.exported_files}")
 """
 
-from .imagen_sdk import (
-    # Main client
-    ImagenClient,
-
-    # Pydantic Models for data structures
-    Profile,
-    ProfileApiResponse,
-    ProfileApiData,
-    ProjectCreationResponseData,
-    ProjectCreationResponse,
-    FileUploadInfo,
-    PresignedUrl,
-    PresignedUrlList,
-    PresignedUrlResponse,
-    EditOptions,
-    StatusDetails,
-    StatusResponse,
-    DownloadLink,
-    DownloadLinksList,
-    DownloadLinksResponse,
-    UploadResult,
-    UploadSummary,
-    QuickEditResult,
-
-    # Exceptions
-    ImagenError,
-    AuthenticationError,
-    ProjectError,
-    UploadError,
-    DownloadError,  # Added this missing export
-
-    # Enums for editing options
-    PhotographyType,
-    CropAspectRatio,
-
-    # Convenience functions
-    quick_edit,
-    get_profiles,
+from .imagen_sdk import ImagenClient, quick_edit, get_profiles
+from .models import (
+    Profile, ProfileApiResponse, ProfileApiData, ProjectCreationResponseData, ProjectCreationResponse,
+    FileUploadInfo, PresignedUrl, PresignedUrlList, PresignedUrlResponse, EditOptions, StatusDetails,
+    StatusResponse, DownloadLink, DownloadLinksList, DownloadLinksResponse, UploadResult, UploadSummary, QuickEditResult
 )
+from .exceptions import ImagenError, AuthenticationError, ProjectError, UploadError, DownloadError
+from .enums import PhotographyType, CropAspectRatio
 
 # Version info
 __version__ = "1.0.0"
@@ -107,18 +76,13 @@ __url__ = "https://github.com/imagenai/imagen-ai-sdk"
 
 # Main exports for public API
 __all__ = [
-    # Main class
     'ImagenClient',
-
-    # Core Pydantic Models (commonly used)
     'Profile',
     'EditOptions',
     'UploadResult',
     'UploadSummary',
     'QuickEditResult',
     'StatusDetails',
-
-    # Response Models (for advanced usage)
     'ProfileApiResponse',
     'ProfileApiData',
     'ProjectCreationResponseData',
@@ -131,22 +95,14 @@ __all__ = [
     'DownloadLink',
     'DownloadLinksList',
     'DownloadLinksResponse',
-
-    # Exceptions
     'ImagenError',
     'AuthenticationError',
     'ProjectError',
     'UploadError',
     'DownloadError',
-
-    # Enums
     'PhotographyType',
     'CropAspectRatio',
-
-    # Functions
     'quick_edit',
     'get_profiles',
-
-    # Metadata
     '__version__',
 ]
